Inline AbstractCommandLine
