+++ hazelcast-command-line/src/test/java/com/hazelcast/commandline/JavaOptionsConsumerTest.java
@@ -49,7 +49,7 @@ public class JavaOptionsConsumerTest {
         Stack<String> args = new Stack<>();
         args.push(options[0] + split + options[1]);
         // when
-        AbstractCommandLine.JavaOptionsConsumer javaOptionsConsumer = new AbstractCommandLine.JavaOptionsConsumer();
+        HazelcastCommandLine.JavaOptionsConsumer javaOptionsConsumer = new HazelcastCommandLine.JavaOptionsConsumer();
         javaOptionsConsumer.consumeParameters(args, argSpec, commandSpec);
         // then
         List<String> list = argSpec.getValue();
@@ -63,7 +63,7 @@ public class JavaOptionsConsumerTest {
         // given
         Stack<String> args = new Stack<>();
         // when
-        AbstractCommandLine.JavaOptionsConsumer javaOptionsConsumer = new AbstractCommandLine.JavaOptionsConsumer();
+        HazelcastCommandLine.JavaOptionsConsumer javaOptionsConsumer = new HazelcastCommandLine.JavaOptionsConsumer();
         javaOptionsConsumer.consumeParameters(args, argSpec, commandSpec);
         // then
         // Exception expected
